Implement list_object_keys_flat function
We are implementing a list_object_keys_flat function which returns a list of
objects instead of a generator. This function should not be used when trying to
fetch a large number of objects. However, one advantage of using this function
is that it is retriable by tenacity

Change-Id: I046d738b676283612ca1a14616d0e2981eac22bd
JIRA: https://helpshift.atlassian.net/browse/OPS-2311